Update UI, fix bugs

+++ app/api-requests.tsx
@@ -8,35 +8,103 @@ import WeaponType from "@/models/weaponType";
 class ApiRequests {
   static hostName = "http://localhost:5161/api/";
 
+  public static async attemptRegister(username: string, password: string): Promise<boolean> {
+    const response = await fetch(this.hostName + "register", {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ username, password }),
+    });
+    return response.ok;
+  }
+
+  public static async attemptLogin(username: string, password: string): Promise<string> {
+    const response = await fetch(this.hostName + "login", {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ username, password }),
+    });
+    if (response.status === 401) {
+      return "Unauthorized";
+    }
+    const data = await response.json();
+    return data.token;
+  }
+
+  public static async getProfile(token: string): Promise<any> {
+    if (!token) {
+      throw new Error("No token found.");
+    }
+    const response = await fetch(this.hostName + "profile", {
+      method: "GET",
+      headers: {
+        "Authorization": `Bearer ${token}`,
+      },
+    });
+    if (!response.ok) {
+      throw new Error("Failed to fetch profile.");
+    }
+    return await response.json();
+  }
+
   public static async fetchPlayerCharacter(id: string): Promise<PlayerCharacter> {
-    const response = await fetch(this.hostName + "playercharacters/" + id);
+    const token = localStorage.getItem("token");
+    if (!token) {
+      throw new Error("No token found.");
+    }
+    const response = await fetch(this.hostName + "playercharacters/" + id, {
+      method: "GET",
+      headers: {
+        "Authorization": `Bearer ${token}`,
+      },
+    });
     const data = await response.json();
     return new PlayerCharacter(data);
   }
 
   public static async savePlayerCharacter(playerCharacter: PlayerCharacter) {
+    const token = localStorage.getItem("token");
+    if (!token) {
+      throw new Error("No token found.");
+    }
     return await fetch(this.hostName + "playercharacters/" + playerCharacter.id, {
       method: 'PUT',
       headers: {
         'Content-Type': 'application/json',
+        "Authorization": `Bearer ${token}`,
       },
       body: JSON.stringify(playerCharacter),
     });
   }
 
   public static async createPlayerCharacter(playerCharacter: PlayerCharacter) {
+    const token = localStorage.getItem("token");
+    if (!token) {
+      throw new Error("No token found.");
+    }
     return await fetch(this.hostName + "playerCharacters", {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
+        "Authorization": `Bearer ${token}`,
       },
       body: JSON.stringify(playerCharacter),
     });
   }
 
   public static async deletePlayerCharacter(id: string) {
+    const token = localStorage.getItem("token");
+    if (!token) {
+      throw new Error("No token found.");
+    }
     return await fetch(this.hostName + "playerCharacters/" + id, {
       method: 'DELETE',
+      headers: {
+        "Authorization": `Bearer ${token}`,
+      },
     });
   }
